Add helper functions & global variables in config.js
