+++ src/js/views/recipeView.js
@@ -88,22 +88,7 @@ class RecipeView {
           <h2 class="heading--2">Recipe ingredients</h2>
           <ul class="recipe__ingredient-list">
           ${this.#data.ingredients
-            .map(ing => {
-              return `
-            <li class="recipe__ingredient">
-            <svg class="recipe__icon">
-              <use href="${icons}#icon-check"></use>
-            </svg>
-            <div class="recipe__quantity">${
-              ing.quantity ? new Fraction(ing.quantity) : ''
-            }</div>
-            <div class="recipe__description">
-              <span class="recipe__unit">${ing.unit}</span>
-              ${ing.description}
-            </div>
-          </li>
-            `;
-            })
+            .map(ing => this.#generateIngredientMarkup(ing))
             .join('')}
           </ul>
         </div>
@@ -129,6 +114,23 @@ class RecipeView {
           </a>
         </div>
     `;
+  }
+
+  #generateIngredientMarkup(ing) {
+    return `
+            <li class="recipe__ingredient">
+            <svg class="recipe__icon">
+              <use href="${icons}#icon-check"></use>
+            </svg>
+            <div class="recipe__quantity">${
+              ing.quantity ? new Fraction(ing.quantity) : ''
+            }</div>
+            <div class="recipe__description">
+              <span class="recipe__unit">${ing.unit}</span>
+              ${ing.description}
+            </div>
+          </li>
+            `;
   }
 }
